requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==21.2.0
Automat==20.2.0
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
botocore==1.23.34
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.1.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
ec2-hibinit-agent==1.0.0
exceptiongroup==1.2.2
h11==0.14.0
hibagent==1.0.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.5.0
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openai==1.45.1
pexpect==4.8.0
Pillow==9.0.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
tqdm==4.66.5
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
